+++ tests/test_helpers/test_bad_name_attribute_use.py
@@ -10,6 +10,8 @@ from __future__ import (
 import sys
 import unittest
 
+import pytest
+
 import dlint
 
 IS_PYTHON_3_5 = sys.version_info >= (3, 5)
@@ -29,6 +31,7 @@ def get_bad_name_attribute_use_implementation(illegal_name_attributes):
 
 class TestBadNameAttributeUse(dlint.test.base.BaseTest):
 
+    @pytest.mark.filterwarnings('ignore::DeprecationWarning:dlint')
     def test_empty_code(self):
         python_node = self.get_ast_node(
             """
@@ -68,6 +71,7 @@ class TestBadNameAttributeUse(dlint.test.base.BaseTest):
 
         assert result == expected
 
+    @pytest.mark.filterwarnings('ignore::DeprecationWarning:dlint')
     def test_bad_name_attributes_basic(self):
         python_node = self.get_ast_node(
             """
@@ -99,6 +103,7 @@ class TestBadNameAttributeUse(dlint.test.base.BaseTest):
 
         assert result == expected
 
+    @pytest.mark.filterwarnings('ignore::DeprecationWarning:dlint')
     def test_bad_name_attributes_nested(self):
         python_node = self.get_ast_node(
             """
@@ -132,6 +137,7 @@ class TestBadNameAttributeUse(dlint.test.base.BaseTest):
 
         assert result == expected
 
+    @pytest.mark.filterwarnings('ignore::DeprecationWarning:dlint')
     @unittest.skipUnless(IS_PYTHON_3_5, 'async statements introduced in Python 3.5')
     def test_bad_name_attributes_async_nested(self):
         python_node = self.get_ast_node(
@@ -166,6 +172,7 @@ class TestBadNameAttributeUse(dlint.test.base.BaseTest):
 
         assert result == expected
 
+    @pytest.mark.filterwarnings('ignore::DeprecationWarning:dlint')
     def test_bad_name_attributes_nested_overwrite(self):
         python_node = self.get_ast_node(
             """
@@ -195,6 +202,7 @@ class TestBadNameAttributeUse(dlint.test.base.BaseTest):
 
         assert result == expected
 
+    @pytest.mark.filterwarnings('ignore::DeprecationWarning:dlint')
     @unittest.skipUnless(IS_PYTHON_3_5, 'async statements introduced in Python 3.5')
     def test_bad_name_attributes_async_nested_overwrite(self):
         python_node = self.get_ast_node(
@@ -225,6 +233,7 @@ class TestBadNameAttributeUse(dlint.test.base.BaseTest):
 
         assert result == expected
 
+    @pytest.mark.filterwarnings('ignore::DeprecationWarning:dlint')
     def test_bad_name_attributes_multiple_findings(self):
         python_node = self.get_ast_node(
             """
@@ -265,6 +274,7 @@ class TestBadNameAttributeUse(dlint.test.base.BaseTest):
 
         assert result == expected
 
+    @pytest.mark.filterwarnings('ignore::DeprecationWarning:dlint')
     def test_bad_name_attributes_multiple_attributes(self):
         python_node = self.get_ast_node(
             """
@@ -307,6 +317,7 @@ class TestBadNameAttributeUse(dlint.test.base.BaseTest):
 
         assert result == expected
 
+    @pytest.mark.filterwarnings('ignore::DeprecationWarning:dlint')
     def test_bad_name_attributes_overwrite(self):
         python_node = self.get_ast_node(
             """
@@ -334,6 +345,7 @@ class TestBadNameAttributeUse(dlint.test.base.BaseTest):
 
         assert result == expected
 
+    @pytest.mark.filterwarnings('ignore::DeprecationWarning:dlint')
     def test_bad_name_attributes_no_module(self):
         python_node = self.get_ast_node(
             """
@@ -365,6 +377,7 @@ class TestBadNameAttributeUse(dlint.test.base.BaseTest):
 
         assert result == expected
 
+    @pytest.mark.filterwarnings('ignore::DeprecationWarning:dlint')
     def test_bad_name_attributes_multiple_attribute_calls(self):
         python_node = self.get_ast_node(
             """
@@ -399,6 +412,68 @@ class TestBadNameAttributeUse(dlint.test.base.BaseTest):
             ),
             dlint.linters.base.Flake8Result(
                 lineno=7,
+                col_offset=11,
+                message=linter._error_tmpl
+            )
+        ]
+
+        assert result == expected
+
+    def test_bad_name_attributes_module_path(self):
+        python_node = self.get_ast_node(
+            """
+            import bar
+
+            def func():
+                obj = bar.Baz()
+                return obj.foo()
+            """
+        )
+
+        linter = get_bad_name_attribute_use_implementation(
+            {
+                'foo': [
+                    'bar.Baz',
+                ],
+            }
+        )
+        linter.visit(python_node)
+
+        result = linter.get_results()
+        expected = [
+            dlint.linters.base.Flake8Result(
+                lineno=6,
+                col_offset=11,
+                message=linter._error_tmpl
+            )
+        ]
+
+        assert result == expected
+
+    def test_bad_name_attributes_module_path_from_import(self):
+        python_node = self.get_ast_node(
+            """
+            from bar import Baz
+
+            def func():
+                obj = Baz()
+                return obj.foo()
+            """
+        )
+
+        linter = get_bad_name_attribute_use_implementation(
+            {
+                'foo': [
+                    'bar.Baz',
+                ],
+            }
+        )
+        linter.visit(python_node)
+
+        result = linter.get_results()
+        expected = [
+            dlint.linters.base.Flake8Result(
+                lineno=6,
                 col_offset=11,
                 message=linter._error_tmpl
             )